+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/CoreModuleConfig.java
@@ -58,7 +58,7 @@ public class CoreModuleConfig extends ModuleConfig {
     /**
      * The threshold of session time. Unit is ms. Default value is 2.5 min.
      */
-    private long storageSessionTimeout = 150_000;
+    private long storageSessionTimeout = 70_000;
     private final List<String> downsampling;
     /**
      * The period of doing data persistence. Unit is second.